+++ packages/react-components/react-toast/stories/Toast/index.stories.tsx
@@ -14,7 +14,7 @@ export { ToastLifecycle } from './ToastLifecycle.stories';
 export { UpdateToast } from './UpdateToast.stories';
 export { MultipleToasters } from './MultipleToasters.stories';
 export { ToasterLimit } from './ToasterLimit.stories';
-export { CustomAnnounce } from './CustomAnnounce.stories';
+export { FocusKeyboardShortcut } from './FocusKeyboardShortcut.stories';
 
 import descriptionMd from './ToastDescription.md';
